Extract useScreencast
